Move ring to self-contained file
